+++ src/MathEval/MiscMathFunc.hh
@@ -18,6 +18,12 @@ limitations under the License.
 #ifndef MISC_MATH_FUNC_HH
 #define MISC_MATH_FUNC_HH
 template <typename DoubleType>
+DoubleType erfc_inv(DoubleType);
+
+template <typename DoubleType>
+DoubleType erf_inv(DoubleType);
+
+template <typename DoubleType>
 DoubleType derfdx(DoubleType);
 
 template <typename DoubleType>
@@ -40,5 +46,5 @@ DoubleType vec_min(DoubleType);
 
 template <typename DoubleType>
 DoubleType dot2dfunc(DoubleType, DoubleType, DoubleType, DoubleType);
-
 #endif
+